WIP: Make full screen use desktop resolution
The ideal would be to have borderless windowed fullscreen but this is
not supported in Irrlicht 1.8.5. It does exist as a patch in minetest
Irrlicht. Until it is properly supported we use full screen with the
current resolution.

The selection of screen resolution is removed from the options panel and
we make the window resizable if it is floating.